requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accent-vc=1.0=dev_0
aom=3.3.0=h27087fc_1
appdirs=1.4.4=pyh9f0ad1d_0
audioread=2.1.9=py310hff52083_3
beautifulsoup4=4.11.1=pyha770c72_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py310hff52083_2
cffi=1.15.0=py310h0fdd8cc_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=37.0.1=py310h9ce1e76_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
ffmpeg=5.0.1=h594f047_1
filelock=3.6.0=pyhd8ed1ab_0
fonttools=4.33.3=py310h5764c6d_0
freetype=2.10.4=h0708190_1
gdown=4.4.0=pyhd8ed1ab_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
icu=70.1=h27087fc_0
idna=3.3=pyhd8ed1ab_0
jbig=2.1=h7f98852_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h166bdaf_1
kiwisolver=1.4.2=py310hbf28c38_1
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=14_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=14_linux64_openblas
libdeflate=1.10=h7f98852_0
libdrm=2.4.109=h7f98852_0
libffi=3.4.2=h7f98852_5
libflac=1.3.4=h27087fc_0
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h516909a_0
liblapack=3.9.0=14_linux64_openblas
libllvm11=11.1.0=hf817b99_3
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpciaccess=0.16=h516909a_0
libpng=1.6.37=h21135ba_2
librosa=0.9.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.3.0=h542a066_3
libuuid=2.32.1=h7f98852_1000
libva=2.14.0=h7f98852_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxml2=2.9.14=h22db469_0
libzlib=1.2.11=h166bdaf_1014
llvmlite=0.38.0=py310h58363a5_1
lz4-c=1.9.3=h9c3ff4c_1
matplotlib-base=3.5.2=py310h5701ce4_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h27087fc_1
nettle=3.6=he412f7d_0
numba=0.55.1=py310ha5257ce_1
numpy=1.21.6=py310h45f3432_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=3.0.3=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pillow=9.1.0=py310he619898_2
pip=22.1=pyhd8ed1ab_0
pooch=1.6.0=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310hff52083_5
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.10.4=h2660328_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
readline=8.1=h46c0cb4_0
requests=2.27.1=pyhd8ed1ab_0
resampy=0.2.2=py_0
scikit-learn=1.1.0=py310hffb9edd_0
scipy=1.8.0=py310hea5193d_1
setuptools=62.2.0=py310hff52083_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.5=h4ff8645_0
svt-av1=0.9.1=h27087fc_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h27826a3_0
torch=1.11.0=pypi_0
tqdm=4.64.0=pyhd8ed1ab_0
typing-extensions=4.2.0=pypi_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py310h5764c6d_1
urllib3=1.26.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
zlib=1.2.11=h166bdaf_1014
zstd=1.5.2=ha95c52a_0
